+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added support for `:async` and `:queue` modes for `reindex` on relation
 - Added basic protection from unfiltered parameters to `where` option
 - Added `models` option to `similar` method
+- Changed async reindex to fetch ids instead of using ranges for numeric primary keys with Active Record
 - Anchor regular expressions by default
 - Raise error when `search` called on relations
 - Raise `ArgumentError` (instead of warning) for invalid regular expression modifiers